Hack to fix build order for py27-oldest tox testenv
The py27-oldest has a specific build order needed (numpy must be built
and installed before matplotlib), so moved the dependencies out of tox.ini
to a requirements file. This relies on the behavior of pip building
dependencies in the order they will be installed, which I'm not sure is
guaranteed.

I also added py27-oldest-prebuild.sh, which is needed for build and
runtime dependencies on Ubuntu 18.

Finally, added some documentation about when the _piecewise()
replacement for numpy.piecewise() is necessary.